==> notebooks/1_data-preparation.ipynb <==
import pandas as pd

wine = pd.read_csv("../data/wine.csv")
auto_mpg = pd.read_csv("../data/auto_mpg.csv")

wine.info()

wine.isnull().sum()

auto_mpg.isnull().sum()

auto_mpg = auto_mpg.dropna()
auto_mpg.isnull().sum()


Было решено удалить строки с пропущенными значениями, так как таких значений всего 6.
Все остальные значения уже конвертированы в числовые (например, origin в auto_mpg).

wine.to_csv("../data/wine_preprocessed.csv")
auto_mpg.to_csv("../data/auto_mpg_preprocessed.csv")
