==> notebooks/2_classification.ipynb <==
import pandas as pd
import numpy as np
from sklearn import linear_model, model_selection, metrics, \
    preprocessing, neighbors, tree, naive_bayes, svm, pipeline

# Загружаем обработанный набор данных
wine_df = pd.read_csv("../data/wine_preprocessed.csv")
wine_df = wine_df.drop(columns="Unnamed: 0")
wine_df.head()

# Отделяем ключевой атрибут
X = wine_df.drop(columns='Cultivar').values
y = wine_df.iloc[:, :1].values
y = y.ravel()

# Разделяем выборку на обучение и тест

X_train, X_test, y_train, y_test = model_selection.train_test_split(X, y)

# оценка модели без подбора параметров
def initial_evaluation(estimator):
    estimator.fit(X_train, y_train)
    y_pred = estimator.predict(X_test)
    # кросс-валидация
    score = model_selection.cross_val_score(estimator, X, y, cv=3)
    print('Confusion matrix:\n', metrics.confusion_matrix(y_test, y_pred),'\n')
    print('Classification report:\n', metrics.classification_report(y_test,  y_pred), '\n')
    print(score.mean())

# подбор параметров
def grid_search_result(estimator, param_grid):
    print('Given parameters: ', param_grid)
    grid_search = model_selection.GridSearchCV(estimator, param_grid, scoring='f1_micro', cv=3)
    grid_search.fit(X, y)
    print('Best params: ', grid_search.best_params_)
    print('F1 score: ', grid_search.best_score_)

# Объявляем StandardScaler для нормализации данных

scaler = preprocessing.StandardScaler()

### Метод k-ближайших соседей (K-Nearest Neighbors)


# https://scikit-learn.org/stable/modules/generated/sklearn.neighbors.KNeighborsClassifier.html
KN_classifier = neighbors.KNeighborsClassifier()
KN_param_grid = dict(scaler=['passthrough', scaler] ,KNC__n_neighbors=[3, 5, 10, 15, 20], KNC__p=[1, 2], KNC__weights=['uniform', 'distance'])
KN_pipe = pipeline.Pipeline(steps=[('scaler', scaler), ('KNC', KN_classifier)])

print('K-Nearest Neighbors:\n')
initial_evaluation(KN_pipe)
grid_search_result(KN_pipe, KN_param_grid)

### Классификатор дерева решений (Decision Tree Classifier)

# https://scikit-learn.org/stable/modules/tree.html
tree_classifier = tree.DecisionTreeClassifier()
tree_pipe = pipeline.Pipeline(steps=[('scaler', scaler), ('tree', tree_classifier)])
tree_param_grid = dict(tree__max_depth=[2, 5, 10, 20, 50, 100, None], tree__criterion=['gini', 'entropy'], tree__min_samples_leaf=[1, 2, 3, 4, 5, 10])
print('Decision Tree Classifier:\n')
initial_evaluation(tree_pipe)
grid_search_result(tree_pipe, tree_param_grid)

### Наивный байесовский классификатор (Naive Bayes)

# https://scikit-learn.org/stable/modules/generated/sklearn.naive_bayes.GaussianNB.html
NB_classifier = naive_bayes.GaussianNB()
NB_pipe = pipeline.Pipeline(steps=[('scaler', scaler), ('bayes', NB_classifier)])
NB_param_grid = dict(bayes__var_smoothing=[1e-09, 1e-12, 1e-05])
print('Naive Bayes:\n')
initial_evaluation(NB_pipe)
grid_search_result(NB_pipe, NB_param_grid)

### Метод опорных векторов (Support Vector Machines)

# https://scikit-learn.org/stable/modules/svm.html
SVM_classifier = svm.SVC()

SVM_pipe = pipeline.Pipeline(steps=[('scaler', scaler), ('SVM', SVM_classifier)])
SVM_param_grid = dict(SVM__C=[0.2, 0.5, 0.7, 1.0, 1.2, 1.5, 1.7, 2, 2.5], SVM__kernel=['linear', 'poly', 'rbf', 'sigmoid'], SVM__tol=[0.1, 0.01, 0.001, 0.0001, 0.00001], SVM__probability=[True, False])
print('Support Vector Machines:\n')
initial_evaluation(SVM_pipe)
grid_search_result(SVM_pipe, SVM_param_grid)

### Логистическая регрессия (Logistic Regression)

# https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.htm
LR_classifier = linear_model.LogisticRegression()
LR_pipe = pipeline.Pipeline(steps=[('scaler', scaler), ('LR', LR_classifier)])
LR_param_grid = dict(LR__C=[0.001, 0.01, 0.1, 0.3, 1, 1.2, 1.5], LR__solver=['liblinear', 'lbfgs', 'newton-cg', 'saga'], LR__tol=[0.1, 0.01, 0.001, 0.0001, 0.00001], LR__max_iter=[1000])
print('Logistic Regression:\n')
initial_evaluation(LR_pipe)
grid_search_result(LR_pipe, LR_param_grid)

### Вывод

Все модели имеют оценку выше, чем 0.9, однако лучше всех отработали Логистическая регрессия и Метод опорных векторов; сильно хуже отработало Дерево решений.
