==> notebooks/3_regression.ipynb <==
import pandas as pd
import numpy as np
from sklearn import linear_model, tree, model_selection, preprocessing, pipeline

# Регрессия

def cross_validation_result(estimator):
    # Проводим кросс-валидацию с разбиением на 3 отрезка
    cv = model_selection.cross_validate(estimator, X, y, scoring=scores, cv=3)
    # Выводим средние результаты
    print('R2 score: ', cv['test_r2'].mean())
    print('MSE score: ', cv['test_neg_mean_squared_error'].mean())
    print('RMSE score: ', cv['test_neg_root_mean_squared_error'].mean())
    print('MAE score: ', cv['test_neg_mean_absolute_error'].mean())


def grid_search_result(estimator, param_grid):
    # Подбираем гиперпараметры с помощью GridSearchCV
    grid = model_selection.GridSearchCV(estimator, param_grid, cv=5)
    grid.fit(X, y)
    cross_validation_result(grid)


### Подготовка данных

# Считываем данные из файла и перемешиваем, так как автомобили отсортированы по году выпуска
mpg = pd.read_csv("../data/auto_mpg_preprocessed.csv")
mpg = mpg.drop(columns="Unnamed: 0").sample(frac=1)

# Отделяем ключевой атрибут и наименования автомобилей 
X = mpg.drop(columns=['mpg','car name']).values
y = mpg['mpg'].values

# Объявляем критерии оценки
scores = ['r2', 'neg_mean_squared_error', 'neg_root_mean_squared_error', 'neg_mean_absolute_error']
# Объявляем скейлер
scaler = preprocessing.StandardScaler()

### Линейная регрессия

%%time
# https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LinearRegression.html
LR = linear_model.LinearRegression()
lr_pipe = pipeline.Pipeline(steps=[('scaler', scaler), ('LR', LR)])

cross_validation_result(lr_pipe)

У линейной регрессии нет гиперпараметров, поэтому подбор не проводится

### Регрессия дерева решений

# https://scikit-learn.org/stable/modules/generated/sklearn.tree.DecisionTreeRegressor.html
tree_regressor = tree.DecisionTreeRegressor()
tree_pipe = pipeline.Pipeline(steps=[('scaler', scaler), ('tree', tree_regressor)])
# Проводим кросс-валидацию
cross_validation_result(tree_pipe)

#### Подбор гиперпараметров для дерева решений
Гиперпараметры дерева решений:
- criterion - функция определения качества разделения вершин
- splitter - стратегия разделения вершин
- max_depth - максимальная глубина дерева
- min_samples_leaf - минимальное кол-во объектов в листе

%%time
# Подбираем гиперпараметры с помощью GridSearchCV
tree_param_grid = dict(tree__criterion=['mse', 'friedman_mse', 'mae', 'poisson'], tree__splitter=['best', 'random'], tree__min_samples_leaf=[1, 3, 5, 10, 15], tree__max_depth=[5, 10, 20, 50, 100, None])
grid_search_result(tree_pipe, tree_param_grid)

### Lasso регрессия

# https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Lasso.html
lasso = linear_model.Lasso()
lasso_pipe = pipeline.Pipeline(steps=[('scaler', scaler), ('lasso', lasso)])
# Проводим кросс-валидацию
cross_validation_result(lasso_pipe)

#### Подбор гиперпараметров для Lasso регрессии
Гиперпараметры lasso:
alpha - штрафующий коэффициент

%%time
# Подбираем гиперпараметры с помощью GridSearchCV
lasso_param_grid = dict(lasso__alpha=np.arange(0.05, 1, 0.05))
grid_search_result(lasso_pipe, lasso_param_grid)

### Гребневая регрессия

# https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Ridge.html
ridge = linear_model.Ridge()
ridge_pipe = pipeline.Pipeline(steps=[('scaler', scaler), ('ridge', ridge)])
# Проводим кросс-валидацию
cross_validation_result(ridge_pipe)

#### Подбор гиперпараметров для Ridge регрессии
Гиперпараметры ridge:
alpha - штрафующий коэффициент

%%time
# Подбираем гиперпараметры с помощью GridSearchCV
ridge_param_grid = dict(ridge__alpha=np.arange(0.05, 1, 0.05))
grid_search_result(ridge_pipe, ridge_param_grid)


### Elastic Net регрессия

# https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.ElasticNet.html
elastic = linear_model.ElasticNet()
elastic_pipe = pipeline.Pipeline(steps=[('scaler', scaler), ('elastic', elastic)])
# Проводим кросс-валидацию
cross_validation_result(elastic_pipe)

#### Подбор гиперпараметров для Elastic Net регрессии
Гиперпараметры Elastic Net:
alpha - штрафующий коэффициент
l1_ratio - коэффициент смешивания, т.е. l1_ratio = 0 будет означать использование только l2 регуляризации, а l1_ratio - только l1

%%time
# Подбираем гиперпараметры с помощью GridSearchCV
elastic_param_grid = dict(elastic__alpha=np.arange(0.05, 1, 0.05), elastic__l1_ratio=np.arange(0.1, 0.9, 0.05))
grid_search_result(elastic_pipe, elastic_param_grid)

### Выводы
В целом, все модели после подбора параметров выдали практически одинаковые результаты. <br>
Скорее всего, это связано с тем, что зависимость ключевого атрибута от большинства параметров является близкой к линейной. <br>
Также, стоит заметить, что для получения таких результатов данные пришлось перемешать, так как они были отсортированы по году выпуска
